repackage get_visuals
